move vote impl into vote mod
