+++ rust/vote-tx-v2/src/gen_tx/vote.rs
@@ -1,0 +1,71 @@
+//! A generalised tx vote struct.
+
+use minicbor::{Decode, Decoder, Encode};
+
+use super::{Choice, Proof, PropId};
+use crate::Cbor;
+
+/// `Vote` array struct length
+const VOTE_LEN: u64 = 3;
+
+/// A vote struct.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Vote<ChoiceT, ProofT, ProopIdT>
+where
+    ChoiceT: for<'a> Cbor<'a>,
+    ProofT: for<'a> Cbor<'a>,
+    ProopIdT: for<'a> Cbor<'a>,
+{
+    /// `choices` field
+    pub(super) choices: Vec<Choice<ChoiceT>>,
+    /// `proof` field
+    pub(super) proof: Proof<ProofT>,
+    /// `prop-id` field
+    pub(super) prop_id: PropId<ProopIdT>,
+}
+
+impl<ChoiceT, ProofT, PropIdT> Decode<'_, ()> for Vote<ChoiceT, ProofT, PropIdT>
+where
+    ChoiceT: for<'a> Cbor<'a>,
+    ProofT: for<'a> Cbor<'a>,
+    PropIdT: for<'a> Cbor<'a>,
+{
+    fn decode(d: &mut Decoder<'_>, (): &mut ()) -> Result<Self, minicbor::decode::Error> {
+        let Some(VOTE_LEN) = d.array()? else {
+            return Err(minicbor::decode::Error::message(format!(
+                "must be a defined sized array with {VOTE_LEN} entries"
+            )));
+        };
+
+        let choices = Vec::<Choice<_>>::decode(d, &mut ())?;
+        if choices.is_empty() {
+            return Err(minicbor::decode::Error::message(
+                "choices array must has at least one entry",
+            ));
+        }
+        let proof = Proof::decode(d, &mut ())?;
+        let prop_id = PropId::decode(d, &mut ())?;
+        Ok(Self {
+            choices,
+            proof,
+            prop_id,
+        })
+    }
+}
+
+impl<ChoiceT, ProofT, PropIdT> Encode<()> for Vote<ChoiceT, ProofT, PropIdT>
+where
+    ChoiceT: for<'a> Cbor<'a>,
+    ProofT: for<'a> Cbor<'a>,
+    PropIdT: for<'a> Cbor<'a>,
+{
+    fn encode<W: minicbor::encode::Write>(
+        &self, e: &mut minicbor::Encoder<W>, (): &mut (),
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        e.array(VOTE_LEN)?;
+        self.choices.encode(e, &mut ())?;
+        self.proof.encode(e, &mut ())?;
+        self.prop_id.encode(e, &mut ())?;
+        Ok(())
+    }
+}
